requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiohttp-proxy==0.1.2
aiosignal==1.2.0
aiosocks==0.2.6
aiosocksy==0.1.2
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.9.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
dateparser==1.1.3
Django==4.1.3
django-scheduler==0.10.0
frozenlist==1.3.1
gino==1.0.1
google-auth==2.14.1
h11==0.14.0
icalendar==5.0.2
idna==3.4
install==1.3.5
multidict==6.0.2
outcome==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PySocks==1.7.1
pyTelegramBotAPI==4.7.1
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
rsa==4.9
schedule==1.1.0
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
telebot==0.0.4
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
undetected-chromedriver==3.1.6
urllib3==1.26.12
websockets==10.4
wsproto==1.2.0
yarl==1.8.1
